Rework domain
This patch completely replaces the previous domain. Now, the domain is
as follows:

A `Budget` is the top-level aggregate. It contains
- Payees
- Categories
- Accounts
- Monthly Budgets

A `Payee` is an external entity with whom transacations are completed.
E.g. "McDonald's" is the payee when you buy a Big Mac.
Likewise (although somewhat confusingly), money is also received from
payees, e.g. "Google" is the payee if you receive a paycheque from them.

A `Category` is an envelope into which available funds are placed (to
earmark those funds) and from which funds are spent.
E.g. You might put $100 into a category called "Dining Out" each month.
Every time you go to a restaurant you mark the resulting transaction
with the category "Dining Out" and the available funds in that
category are reduced by the amount of the transaction.

An `Account` represents anything that contains your money or your debts.
In other words, if it has a balance it's probably an account, e.g. your
bank accounts, credit cards, and cash (don't forget to track your
cash!).
An `Account` contains `Transaction`s.

A `Transaction` is an exchange of money that changes the balance of an
account. Transactions are typically completed with an external Payee,
but can also be transfers from one account to another.

A `Monthly Budget` contains the state of the budget for a given month.
Each month, fund are assigned to categories and expense transactions
spend money from the categories. Leftover funds in each category are
rolled over to the next month.

A `Monthly Category` contains the state of the reference `Category` for
the given month, for the same reasons as above.